Refactored non cli commands into core.go. Removed redundant tests.
